try decode plaintext keyring first
if we try to decode keyring as binary encoded blob, the code throws.
and this is expected, as the exception will be caught. but somehow,
in KeyRing::decode(), the buffer::end_of_buffer is not caught. hence
we will have when running `ceph` CLI with a human-readable keyring:

libc++abi.dylib: terminating with uncaught exception of type
ceph::buffer::end_of_buffer: buffer::end_of_buffer
Abort trap: 6

to workaround this. we try to decode keyring as plain text. this should
work fine. as the only source of keyring is the configured keyring file,
so we should be able to decode the keyring with this change.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>