fix the deprecated bottle block syntax
Warning: Calling `sha256 "digest" => :tag` in a bottle block is deprecated! Use `brew style --fix` on the formula to update the style or use `sha256 tag: "digest"` instead.
Please report this issue to the tchaikov/ceph-client tap (not Homebrew/brew or Homebrew/core), or even better, submit a PR to fix it:
  /usr/local/Homebrew/Library/Taps/tchaikov/homebrew-ceph-client/ceph-client.rb:10

Signed-off-by: Kefu Chai <tchaikov@gmail.com>