do not override CC or LDSHARED
to preserve the default options, otherwise we have

ld: library not found for -lSystem

as "-isysroot /Library/Developer/CommandLineTools/SDKs/MacOSX.sdk" would
be removed from the setting otherwise.

and the linker would need to find the libpython library when linking the
cythonized extension, and the "clang -bundle -undefined dynamic_lookup"
is overriden.

Signed-off-by: Kefu Chai <tchaikov@gmail.com>